feat: rootless container support
attempts to autodetect if the container runtime is rootless, if it
cannot be detected then will assume rootful for docker and rootless for
podman as these are the most common way to install.

in addition environment variable PREK_CONTAINER_RUNTIME_ROOTLESS has
also been provided to allow detection to be overridden if required, this
variable should be treated as a boolean.

Signed-off-by: Steven Taylor <steven@taylormuff.co.uk>